Move headingRank beside touchTarget, which it argues from
`REGION_HEADING` is about to be used by `/art` and `SessionSchedule`, neither of
which is a conditions component, so a constant living under `conditions/` stops
describing where it belongs.

`ui/` rather than `components/`. Its own docstring argues from `touchTarget.ts`
-- "a rank asserted in two places is one of them waiting to drift, which is the
argument `touchTarget.ts` already makes about a number" -- and that file is the
repo's precedent for this exact kind of module: a named constant shared across
feature folders, in `components/ui/`. `components/` itself holds page-level
components and two feature helpers, which this is not.

ADR-0014 records the path as `src/components/headingRank.ts`, which was never
where the file was. Correcting the ADR is left to the slice that amends it for
the conversions themselves, so this commit moves code and nothing else.

Imports only. No behaviour change, and no test changes beyond the two suites
that import the constant to assert against it.
